Move the tasks/utils.py tests to test_utils.py
