+++ tests/test_workers/test_tasks/test_build.py
@@ -1,5 +1,4 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
-import logging
 import os
 import re
 import textwrap
@@ -8,7 +7,7 @@ from unittest import mock
 import pytest
 
 from iib.exceptions import IIBError
-from iib.workers.tasks import build, utils
+from iib.workers.tasks import build
 
 
 @mock.patch('iib.workers.tasks.build.run_cmd')
@@ -272,21 +271,6 @@ def test_push_arch_image(mock_run_cmd, mock_glps):
     assert 'docker://registry:8443/operator-registry-index:3-amd64' in push_args
 
 
-@pytest.mark.parametrize('use_creds', (True, False))
-@mock.patch('iib.workers.tasks.utils.run_cmd')
-def test_skopeo_inspect(mock_run_cmd, use_creds):
-    mock_run_cmd.return_value = '{"Name": "some-image"}'
-    image = 'docker://some-image:latest'
-    rv = utils.skopeo_inspect(image, use_creds=use_creds)
-    assert rv == {"Name": "some-image"}
-    skopeo_args = mock_run_cmd.call_args[0][0]
-    expected = ['skopeo', 'inspect', image]
-    if use_creds:
-        expected += ['--creds', 'iib:iibpassword']
-
-    assert skopeo_args == expected
-
-
 @pytest.mark.parametrize('request_succeeded', (True, False))
 @mock.patch('iib.workers.tasks.build._prepare_request_for_build')
 @mock.patch('iib.workers.tasks.build.opm_index_add')
@@ -399,40 +383,6 @@ def test_opm_index_add_already_failed(mock_bi, mock_srs, mock_gr):
     mock_srs.assert_called_once()
     mock_gr.assert_called_once_with(3)
     mock_bi.assert_not_called()
-
-
-@mock.patch('iib.workers.tasks.utils.subprocess.run')
-def test_run_cmd(mock_sub_run):
-    mock_rv = mock.Mock()
-    mock_rv.returncode = 0
-    mock_sub_run.return_value = mock_rv
-
-    utils.run_cmd(['echo', 'hello world'], {'cwd': '/some/path'})
-
-    mock_sub_run.assert_called_once()
-
-
-@pytest.mark.parametrize('exc_msg', (None, 'Houston, we have a problem!'))
-@mock.patch('iib.workers.tasks.utils.subprocess.run')
-def test_run_cmd_failed(mock_sub_run, caplog, exc_msg):
-    # When running tests that involve Flask before this test, the iib.workers loggers
-    # are disabled. This is an ugly workaround.
-    for logger in ('iib.workers', 'iib.workers.tasks', 'iib.workers.tasks.utils'):
-        logging.getLogger(logger).disabled = False
-
-    mock_rv = mock.Mock()
-    mock_rv.returncode = 1
-    mock_rv.stderr = 'some failure'
-    mock_sub_run.return_value = mock_rv
-
-    expected_exc = exc_msg or 'An unexpected error occurred'
-    with pytest.raises(IIBError, match=expected_exc):
-        utils.run_cmd(['echo', 'iib:iibpassword'], exc_msg=exc_msg)
-
-    mock_sub_run.assert_called_once()
-    # Verify that the password is not logged
-    assert '********' in caplog.text
-    assert 'iib:iibpassword' not in caplog.text
 
 
 @pytest.mark.parametrize('request_succeeded', (True, False))
